pyproject.toml
[tool.poetry]
name = "movie-booking-be"
version = "0.1.0"
description = "backend server for csci 6050 movie booking app project. Team 5"
authors = ["davidkan <davidkan717@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.43"
flask-sqlalchemy = "^3.1.1"
flask = "^3.1.2"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
flask-cors = "^6.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
